+++ yeyo/cli.py
@@ -2,16 +2,17 @@
 # All Rights Reserved
 
 import functools
+import json
 from pathlib import Path
 
-import py
 import click
-from semver import parse_version_info
+import docker as dockerpy
+import py
 from jinja2 import Template
+from semver import parse_version_info
 
 from yeyo import __version__
 from yeyo.config import YeyoConfig
-
 
 STARTING_VERSION = "0.0.0-dev.1"
 STARTING_FILE = Path("VERSION")
@@ -71,10 +72,57 @@ def version():
     click.echo(__version__)
 
 
-@main.command()
+@main.group()
+def dev():
+    """Entrypoint for dev related commands."""
+
+
+@dev.command()
 def test():
-    """Run the tests."""
+    """Run yeyo's tests through pytest."""
     py.test.cmdline.main(["yeyo"])
+
+
+@dev.group()
+@click.option("-i", "--image", default="thauck/yeyo", help="The docker image name to make.")
+@click.option(
+    "-t",
+    "--tags",
+    default=[__version__, "latest"],
+    help="The tags to ascribe to the image.",
+    multiple=True,
+)
+@click.option("-u", "--username", help="The username to use to authenticate to docker.")
+@click.option("-p", "--password", help="The password to use to authenticate to docker.")
+@click.pass_context
+def docker(ctx, image, tags, username, password):
+    """A docker group with commands for working with yeyo's docker image."""
+
+    ctx.obj["client"] = dockerpy.from_env()
+    ctx.obj["client"].login(username, password)
+
+    ctx.obj["image"] = image
+    ctx.obj["tags"] = [f"{image}:{t}" for t in tags]
+
+
+@docker.command()
+@click.pass_context
+def build(ctx):
+    """Build the docker image with the appropriate tags."""
+
+    image, _ = ctx.obj["client"].images.build(
+        path=".", tag=ctx.obj["tags"], dockerfile="Dockerfile"
+    )
+    click.echo(image)
+
+
+@docker.command()
+@click.pass_context
+def push(ctx):
+    """Push the docker image to docker hub."""
+
+    push = ctx.obj["client"].images.push(ctx.obj["image"])
+    print(json.dumps(push, indent=2).replace(r"\r\n", "\n"))
 
 
 @main.command()